Clearing some code from the Hackathon and adjusting front
